Sprint 4 almost done

+++ src/modules/content/services/DeleteContentService.ts
@@ -2,13 +2,9 @@ import { inject, injectable } from 'tsyringe';
 
 import { Content } from '@prisma/client';
 
-// import AppError from '@shared/errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 import IContentRepository from '../repositories/IContentRepository';
-
-interface IRequest {
-  id: string;
-}
 
 @injectable()
 export default class DeleteContentService {
@@ -17,9 +13,11 @@ export default class DeleteContentService {
     private contentRepository: IContentRepository,
   ) { }
 
-  public async execute({
-    id,
-  }: IRequest): Promise<Content> {
+  public async execute(id: string): Promise<Content> {
+    const contentExists = await this.contentRepository.findByID(id);
+
+    if (!contentExists) throw new AppError('A content with this id does not exist');
+
     const content = await this.contentRepository.delete(id);
 
     return content;
